Move scan file operations to `scan_file_dialog.rs`
